Bugfix gradient

+++ rec.py
@@ -167,8 +167,8 @@ def _plot_labels(data_path: Path) -> None:
     _label_path(data_path)
     y_trues = np.load(str(data_path / "sample_y.npy"))
     y_preds = (np.load(str(data_path / "sample_as.npy")) > 1).astype(np.uint8)
-    y_rc = (np.load(str(data_path / "sample_rc.npy"))).astype(np.uint8)
-    y_oc = (np.load(str(data_path / "sample_oc.npy"))).astype(np.uint8)
+    y_rc = (np.load(str(data_path / "sample_rc.npy")))
+    y_oc = (np.load(str(data_path / "sample_oc.npy")))
 
     col_len, col_sep = 10, 5
     row_len, row_sep = 10, 50
